backend/requirements.txt
aiofiles==24.1.0
annotated-doc==0.0.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
audioldm==0.1.1
audioread==3.1.0
auraloss==0.4.0
blis==1.3.0
brotli==1.2.0
catalogue==2.0.10
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
cloudpathlib==0.23.0
cloudpickle==3.1.2
colorama==0.4.6
colorlog==6.10.1
confection==0.1.5
cymem==2.0.11
decorator==5.2.1
deepfilternet==0.5.6
demucs==4.0.1
docopt==0.6.2
dora_search==0.1.12
einops==0.8.1
exceptiongroup==1.3.0
fastapi==0.121.0
ffmpy==0.6.4
filelock==3.20.0
flashy==0.0.2
fsspec==2025.10.0
ftfy==6.3.1
gradio==5.49.1
gradio_client==1.13.3
groovy==0.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
julius==0.2.7
lameenc==1.8.1
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.9.2
lightning-utilities==0.15.2
llvmlite==0.45.1
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
murmurhash==1.0.13
networkx==3.4.2
num2words==0.5.14
numba==0.62.1
numpy==1.23.5
omegaconf==2.3.0
openunmix==1.3.0
orjson==3.11.4
packaging==25.0
pandas==2.3.3
pillow==11.3.0
platformdirs==4.5.0
pooch==1.8.2
preshed==3.0.10
progressbar==2.5
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
resampy==0.4.3
retrying==1.4.2
rich==14.2.0
ruff==0.14.4
safehttpx==0.1.7
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
semantic-version==2.10.0
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
smart_open==7.4.4
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.49.3
submitit==1.5.3
sympy==1.14.0
thinc==8.3.8
threadpoolctl==3.6.0
tokenizers==0.22.1
tomlkit==0.13.3

# FIXED VERSIONS BELOW
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
xformers==0.0.27.post2

torchdiffeq==0.2.5
torchlibrosa==0.0.9
torchmetrics==1.8.2
tqdm==4.67.1
transformers==4.46.0

treetable==0.2.6
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
wasabi==1.1.3
wcwidth==0.2.14
weasel==0.4.2
websockets==15.0.1
wrapt==2.0.1

# Security & Rate Limiting
python-dotenv==1.0.1
slowapi==0.1.9
google-generativeai==0.8.6
beautifulsoup4==4.14.3
youtube-transcript-api==1.2.3

